Refactor internal functions to use opts map
Many internal functions had two or three arguments for customizing
behavior. Since we'd like to be able to add more customization in the
future, condensing these into an options map seems like a good move.

One subtle difference introduced is that the functions which previously
read a :date-format value from a map and used `or` to determine whether
to used the default-date-format, now use `merge` and so have different
behavior if the `:date-format` key is present but `nil`. I'm assuming
this will be fine.